+++ tests/Model/YamlSpacesBetweenGroups/YamlSpacesBetweenGroupsCheckerTest.php
@@ -56,6 +56,6 @@ class YamlSpacesBetweenGroupsCheckerTest extends TestCase
      */
     private function getStandardsParametersData(int $level)
     {
-        return new StandardParametersData(4, 4, $level);
+        return new StandardParametersData(4, 4, $level, null);
     }
 }