+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 - whoever can create own checker and fixer and add it to config file
 - create new standards: YamlEmptyLineAtEnd
     - Check yaml file has empty line at end of file
+- create new standards: YamlServiceAliasing
+    - Check yaml service file observe short or long code style aliasing
 
 ### Changed
 - remove command options and move it to config file